backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
pypdf>=4.0.0
faiss-cpu>=1.7.4
sentence-transformers>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
google-generativeai>=0.4.0